requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0

boto3~=1.20.22
botocore~=1.23.22